+++ backend/fittoring/src/main/java/fittoring/FittoringApplication.java
@@ -1,4 +1,4 @@
-package com.fittoring.fittoring;
+package fittoring;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class FittoringApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(FittoringApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(FittoringApplication.class, args);
+    }
 
 }
