+++ backend/fittoring/src/main/java/fittoring/mentoring/CategoryMentoring.java
@@ -1,6 +1,5 @@
-package com.fittoring.fittoring;
+package fittoring.mentoring;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -8,31 +7,23 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "reservation")
+@Table(name = "category_mentoring")
 @Entity
-public class Reservation {
+public class CategoryMentoring {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false)
-    private String menteeName;
-
-    @Column(nullable = false, unique = true)
-    private String menteePhone;
-
-    private String context;
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
     @ManyToOne
     @JoinColumn(nullable = false)
     private Mentoring mentoring;
+
+    @ManyToOne
+    @JoinColumn(nullable = false)
+    private Category category;
 }
