+++ backend/fittoring/src/main/java/fittoring/mentoring/CategoryMentoringRepository.java
@@ -1,4 +1,4 @@
-package com.fittoring.fittoring;
+package fittoring.mentoring;
 
 import org.springframework.data.repository.ListCrudRepository;
