+++ backend/fittoring/src/main/java/fittoring/mentoring/CategoryRepository.java
@@ -1,8 +1,9 @@
-package com.fittoring.fittoring;
+package fittoring.mentoring;
 
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OfferingRepository extends ListCrudRepository<Offering, Long> {
+public interface CategoryRepository extends ListCrudRepository<Category, Long> {
+
 }
