+++ backend/fittoring/src/main/java/fittoring/mentoring/ImageRepository.java
@@ -1,9 +1,8 @@
-package com.fittoring.fittoring;
+package fittoring.mentoring;
 
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MentoringRepository extends ListCrudRepository<Mentoring, Long> {
-
+public interface ImageRepository extends ListCrudRepository<Image, Long> {
 }
