+++ backend/fittoring/src/main/java/fittoring/mentoring/Mentoring.java
@@ -1,9 +1,7 @@
-package com.fittoring.fittoring;
+package fittoring.mentoring;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -12,21 +10,27 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "image")
+@Table(name = "mentoring")
 @Entity
-public class Image {
+public class Mentoring {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(columnDefinition = "TEXT", nullable = false)
-    private String url;
-
-    @Enumerated(EnumType.STRING)
     @Column(nullable = false)
-    private ImageType relationType;
+    private String mentorName;
+
+    @Column(nullable = false, unique = true)
+    private String mentorPhone;
 
     @Column(nullable = false)
-    private Long relationId;
+    private int price;
+
+    private Integer career;
+
+    private String content;
+
+    @Column(nullable = false)
+    private String introduction;
 }
