+++ backend/fittoring/src/main/java/fittoring/mentoring/OfferingRepository.java
@@ -1,9 +1,8 @@
-package com.fittoring.fittoring;
+package fittoring.mentoring;
 
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends ListCrudRepository<Category, Long> {
-
+public interface OfferingRepository extends ListCrudRepository<Offering, Long> {
 }
