+++ backend/fittoring/src/main/java/fittoring/mentoring/Reservation.java
@@ -1,23 +1,38 @@
-package com.fittoring.fittoring;
+package fittoring.mentoring;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "category")
+@Table(name = "reservation")
 @Entity
-public class Category {
+public class Reservation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Column(nullable = false)
+    private String menteeName;
+
     @Column(nullable = false, unique = true)
-    private String title;
+    private String menteePhone;
+
+    private String context;
+
+    @Column(nullable = false)
+    private LocalDateTime createdAt;
+
+    @ManyToOne
+    @JoinColumn(nullable = false)
+    private Mentoring mentoring;
 }
