+++ backend/fittoring/src/test/java/fittoring/FittoringApplicationTests.java
@@ -1,4 +1,4 @@
-package com.fittoring.fittoring;
+package fittoring;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class FittoringApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
